+++ test/utils_shared.py
@@ -615,6 +615,10 @@ def prepare_spec_arguments(spec, arguments, opname, entity_map, with_txn_callbac
         # Aggregate uses "batchSize", while find uses batch_size.
         elif (arg_name == "batchSize" or arg_name == "allowDiskUse") and opname == "aggregate":
             continue
+        elif arg_name == "bypassDocumentValidation" and (
+            opname == "aggregate" or "find_one_and" in opname
+        ):
+            continue
         elif arg_name == "timeoutMode":
             raise unittest.SkipTest("PyMongo does not support timeoutMode")
         # Requires boolean returnDocument.